refactor(runtimes): split provider files to slim shim binary
Move heavy install/list/migrate methods (Install, Uninstall, ListInstalled,
ListAvailable, DetectInstalled, GlobalPackages, InstallGlobalPackages,
ManualPackageInstallCommand, plus install helpers) and the version
getters/setters into provider_full.go files tagged //go:build !shim.
Provider remains an alias for ShimProvider in shim builds via the new
provider_shim.go, so the registry and its callers don't change. The
Node.js lifecycle helper and runtime test harness are also tagged !shim.

Build the shim with -tags shim (rnr.yaml + both CI workflows) so the
linker drops net/http, klauspost/compress, sevenzip, brotli, lz4, xz,
bzip2, archive/zip, crypto/tls, afero, hashicorp/golang-lru,
progressbar, the embedded manifest JSON, and the Node release schedule.
Shim drops from 5.70 MB to 2.68 MB (-53%); main CLI is unchanged.
