+++ .github/workflows/release.yml
@@ -187,7 +187,7 @@ jobs:
 
     - name: Build shim executable
       run: |
-        go build -v -ldflags="-s -w" -trimpath -o dist/dtvem-shim${{ matrix.goos == 'windows' && '.exe' || '' }} ./src/cmd/shim
+        go build -v -ldflags="-s -w" -trimpath -tags shim -o dist/dtvem-shim${{ matrix.goos == 'windows' && '.exe' || '' }} ./src/cmd/shim
       shell: bash
       env:
         GOOS: ${{ matrix.goos }}
