+++ src/internal/runtime/provider_test_harness.go
@@ -1,3 +1,5 @@
+//go:build !shim
+
 package runtime
 
 import (
